backend/pyproject.toml
[project]
name = "react-chat-ui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "uvicorn[standard]>=0.35.0",
    "cloudscraper>=1.2.71",
    "ddgs>=9.6.0",
    "fake-useragent>=2.2.0",
    "fastapi[standard]>=0.116.2",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-core>=0.3.76",
    "langchain-experimental>=0.3.4",
    "langchain-ollama>=0.3.7",
    "langchain-postgres>=0.0.15",
    "langgraph>=0.6.7",
    "langgraph-checkpoint-postgres>=2.0.23",
    "langgraph-cli[inmem]>=0.4.2",
    "pandas>=2.3.2",
    "psycopg[binary,pool]>=3.2.10",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "selenium>=4.35.0",
    "tabulate>=0.9.0",
    "undetected-chromedriver>=3.5.5",
    "ollama>=0.5.3",
    "beautifulsoup4>=4.14.0",
    "odfpy>=1.4.1",
    "playwright>=1.55.0",
    "webdriver-manager>=4.0.2",
    "pyinstaller>=6.16.0",
]
